# Replace go-bindata with //go:embed