--- a/staging/src/k8s.io/kubectl/pkg/util/i18n/i18n.go
+++ b/staging/src/k8s.io/kubectl/pkg/util/i18n/i18n.go
@@ -19,16 +19,18 @@ package i18n
 import (
 	"archive/zip"
 	"bytes"
+	"embed"
 	"errors"
 	"fmt"
 	"os"
 	"strings"
 
-	"k8s.io/kubectl/pkg/generated"
-
 	"github.com/chai2010/gettext-go/gettext"
 	"k8s.io/klog/v2"
 )
+
+//go:embed translations
+var translations embed.FS
 
 var knownTranslations = map[string][]string{
 	"kubectl": {
@@ -112,7 +114,7 @@ func LoadTranslations(root string, getLanguageFn func() string) error {
 		if err != nil {
 			return err
 		}
-		data, err := generated.Asset(filename)
+		data, err := translations.ReadFile(filename)
 		if err != nil {
 			return err
 		}