--- a/test/e2e/framework/testfiles/testdata/a/foo.txt
+++ b/test/e2e/framework/testfiles/testdata/a/foo.txt
@@ -1,0 +1,1 @@
+Hello World