#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

With kubernetes/kubernetes using Go 1.16 (ref: https://github.com/kubernetes/kubernetes/pull/98572), we should be exploring at better ways of packaging/bundling binary data.

The package that we currently use for embedding binary data (go-bindata) is largely unmaintained and we have had to update references to the package (ref: https://github.com/kubernetes/kubernetes/issues/96169). Go 1.16 has built-in support for embedding files in the compiled binaries.

This pull request explores removing the dependency on go-bindata. Currently, we use go-bindata for packaging the following:
- kubectl translations data
- Conformance test fixtures

##### Tasks

- [x] Migrate kubectl translation data
- [x] Migrate conformance text fixtures
- [x] Remove go-bindata related artifacts

#### Which issue(s) this PR fixes:

Fixes #99150

#### Special notes for your reviewer:

##### kubectl translations
 
Go 1.16's embed directive doesn't allow embedding files from parent directories. Hence, moving the translations data to inside the i18n package.

Logically speaking as well, kubectl related artifacts should be inside the kubectl package. I think historically the files were at the root of k/k because of the generate script.

Comparing compilation times and binary size for kubectl:

```
(on this branch)
❯ time make WHAT=./cmd/kubectl
...
make WHAT=./cmd/kubectl  336.24s user 31.96s system 315% cpu 1:56.56 total
❯ ls -l _output/local/bin/linux/amd64/kubectl
-rwxr-xr-x 1 nabarun nabarun 46346240 Jun  4 10:02 _output/local/bin/linux/amd64/kubectl

(on master)
❯ time make WHAT=./cmd/kubectl
...
make WHAT=./cmd/kubectl  348.77s user 39.83s system 318% cpu 2:01.97 total
❯ ls -l _output/local/bin/linux/amd64/kubectl
-rwxr-xr-x 1 nabarun nabarun 46366720 Jun  4 07:23 _output/local/bin/linux/amd64/kubectl
```

##### Conformance test fixtures

The conformance test fixtures are spread over several packages and hence not as trivial to embed. 

	"test/conformance/testdata/..."
	"test/e2e/testing-manifests/..."
	"test/e2e_node/testing-manifests/..."
	"test/images/..."
	"test/fixtures/..."

Inside each of the above directories, there is a function which needs to be called whenever those fixtures/testdata need to be used. 

##### References

https://golang.org/pkg/embed/
https://go.googlesource.com/proposal/+/master/design/draft-embed.md

#### Does this PR introduce a user-facing change?

```release-note
Replace go-bindata with //go:embed 
```

#### Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.:

```docs
NONE
```

/assign
/priority important-soon

/sig architecture
/cc @dims

/sig cli
/area kubectl
/cc @soltysh