@soltysh 2021-03-05
@palnabarun wow this looks great! the translation should've been moved along with the rest of kubectl last release, thanks!