@BenTheElder 2021-06-03
- How does this affect compile times, now that instead of statically generated files we include sources using the go tool at build time?
- How does this affect library users (thinking kubectl)? At minimum they must need new enough go to support this? Does the go module server / dependency management handle this correctly? (like the source export zips would now need to have the source files instead of the generated go files)?