@palnabarun 2021-06-04
> How does this affect compile times, now that instead of statically generated files we include sources using the go tool at build time?

I tried on just one platform (linux/amd64) and the compile time was similar in both cases. It was a bit faster with the changes here, but can be random behavior as well. I am yet to test the same on other platforms. I need to dig deeper into our release tooling to see how we compile our user-facing binaries.

The binary sizes are also similar, with the binary without embedded files a bit smaller.
```
(on this branch)
❯ time make WHAT=./cmd/kubectl
...
make WHAT=./cmd/kubectl  336.24s user 31.96s system 315% cpu 1:56.56 total
❯ ls -l _output/local/bin/linux/amd64/kubectl
-rwxr-xr-x 1 nabarun nabarun 46346240 Jun  4 10:02 _output/local/bin/linux/amd64/kubectl
```

```
(on master)
❯ time make WHAT=./cmd/kubectl
...
make WHAT=./cmd/kubectl  348.77s user 39.83s system 318% cpu 2:01.97 total
❯ ls -l _output/local/bin/linux/amd64/kubectl
-rwxr-xr-x 1 nabarun nabarun 46366720 Jun  4 07:23 _output/local/bin/linux/amd64/kubectl
```

The full log is [here](https://gist.github.com/palnabarun/35dcb05f415eccb527a89844bdbd7a07)

> How does this affect library users (thinking kubectl)? At minimum they must need new enough go to support this?

If the users are using the compiled binary that we publish, I don't see any change in behavior. If they are pulling in source code and building things (like kubectl) on their own, they must have Go 1.16, which is a sane ask since k/k depends on Go 1.16.

The same goes for when someone is importing `k8s.io/kubectl`, the package is importable in projects using a Go version supporting the embed package. Otherwise, they would get an error like the following:

```
❯ go1.15 run main2.go
../../kubernetes/kubectl/pkg/util/i18n/i18n.go:22:2: package embed is not in GOROOT (/home/nabarun/sdk/go1.15/src/embed)
```

>  Does the go module server / dependency management handle this correctly? (like the source export zips would now need to have the source files instead of the generated go files)?

I have to explore more this. If the module server treats go source files and other filetypes differently, there _maybe_ an issue.