@palnabarun 2021-06-04
> (like the source export zips would now need to have the source files instead of the generated go files)?

@BenTheElder --  do you have any leads on how I can test if anything changed wrt source export zips locally?

I am reading https://golang.org/ref/mod#module-proxy and https://golang.org/ref/mod#zip-files to get some thoughts.