@BenTheElder 2021-06-04
> @BenTheElder -- do you have any leads on how I can test if anything changed wrt source export zips locally?

I'm not sure. I don't know what all the zips trim if anything either, but the history of go dependency management involves a lot of "trimming non-source files", so just a sanity check 😅  I _imagine_ go handles this as expected, but a quick glance didn't uncover anything, and it stuck out to me as a potential difference versus generated checked-in source files.