@palnabarun 2021-06-09
/retitle Replace go-bindata with //go:embed