@palnabarun 2021-06-09
> > @BenTheElder -- do you have any leads on how I can test if anything changed wrt source export zips locally?
> 
> I'm not sure. I don't know what all the zips trim if anything either, but the history of go dependency management involves a lot of "trimming non-source files", so just a sanity check sweat_smile I _imagine_ go handles this as expected, but a quick glance didn't uncover anything, and it stuck out to me as a potential difference versus generated checked-in source files.

I tried this by pushing [a tag](https://github.com/palnabarun/kubectl/tree/v0.99.0-nabarun.1) in my fork with the translation files.

A `go mod download` of that version of the source code had all the non-source files as well in the go mod cache. I feel this should give a good signal that the non-source files (translations, test fixtures etc.) will be handled correctly.