@dims 2021-06-09
@palnabarun can you please add binary sizes before and with this PR?